Refactor metadata download
